% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{ajout_logo_sgdf}
\alias{ajout_logo_sgdf}
\title{Ajout du logo SGDF}
\usage{
ajout_logo_sgdf(plot = ggplot2::last_plot(), type = "instit")
}
\arguments{
\item{plot}{Un graphique \code{ggplot2}}

\item{type}{La couleur du logo, parmi :
("instit"/"symbole"/"lj"/"sg"/"pk"/"compa"/"farfa"/"vdl")}
}
\value{
Un graphique \code{ggplot2}
}
\description{
Ajout du logo SGDF
}
