% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{check_fonts_in_r}
\alias{check_fonts_in_r}
\title{Verif de la presence des polices et chargement}
\usage{
check_fonts_in_r(fonts = c("Sarabun", "Raleway", "Caveat Brush"))
}
\arguments{
\item{fonts}{Liste des polices}
}
\description{
Verif de la presence des polices et chargement
}
