% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_attach.R
\name{.onAttach}
\alias{.onAttach}
\title{Ajout des polices au chargement du package}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Dummy}

\item{pkgname}{Dummy}
}
\value{
Rien
}
\description{
Ajout des polices au chargement du package
}
