% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{html-dependencies}
\alias{html-dependencies}
\alias{html_dependency_sarabun}
\alias{html_dependency_raleway}
\alias{html_dependency_caveat_brush}
\title{Dependances HTML pour les polices SGDF}
\usage{
html_dependency_sarabun()

html_dependency_raleway()

html_dependency_caveat_brush()
}
\value{
Un objet a inclure dans une liste de dependances de
\link[htmltools:attachDependencies]{attachDependencies}.
}
\description{
\code{html_dependency_sarabun()} fournit la dependance pour Sarabun
\code{html_dependency_raleway()} fournit la dependance pour Raleway
\code{html_dependency_caveat_brush()} fournit la dependance pour Caveat Brush
}
