% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{passage_flex_img}
\alias{passage_flex_img}
\title{Conversion d'un tableau flextable en ggplot}
\usage{
passage_flex_img(flex_table)
}
\arguments{
\item{flex_table}{Un tableau fait avec flextable()}
}
\value{
Un graph ggplot
}
\description{
Conversion d'un tableau flextable en ggplot
}
