% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picto.R
\name{picto}
\alias{picto}
\title{Fournit un picto d'une couleur}
\usage{
picto(picto, couleur = "bleu_instit", scale = "100")
}
\arguments{
\item{picto}{Nom du picto}

\item{couleur}{Couleur
(parmi "bleu_instit"/"orange_lj"/"bleu_sg"/"rouge_pk"/
"vert_compa"/"vert_farfa"/"violet_vdl"/"blanc")}

\item{scale}{Echelle, en px}
}
\value{
Une image
}
\description{
Fournit un picto d'une couleur
}
