% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_color_sgdf_discrete}
\alias{scale_color_sgdf_discrete}
\title{Discrete color scale}
\usage{
scale_color_sgdf_discrete(palette = "pal_sgdf_main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{L'une des palettes de \link{sgdf_palettes}.}

\item{reverse}{Indique si il faut inverser la palette.}

\item{...}{Arguments de \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
Discrete color scale
}
