% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_sgdf_continuous}
\alias{scale_fill_sgdf_continuous}
\title{Continuous fill scale}
\usage{
scale_fill_sgdf_continuous(
  palette = "pal_sgdf_bleu_instit",
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{L'une des palettes de \link{sgdf_palettes}.}

\item{reverse}{Indique si il faut inverser la palette.}

\item{...}{Arguments de \code{\link[ggplot2:scale_gradient]{ggplot2::scale_fill_gradientn()}}.}
}
\description{
Continuous fill scale
}
