% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_sgdf_discrete}
\alias{scale_fill_sgdf_discrete}
\title{Discrete fill scale}
\usage{
scale_fill_sgdf_discrete(palette = "pal_sgdf_main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{L'une des palettes de \link{sgdf_palettes}.}

\item{reverse}{Indique si il faut inverser la palette.}

\item{...}{
  Arguments passed on to \code{\link[=scale_color_sgdf_discrete]{scale_color_sgdf_discrete}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Discrete fill scale
}
