% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{sgdf_colors}
\alias{sgdf_colors}
\title{Vecteur de couleurs}
\usage{
sgdf_colors(...)
}
\arguments{
\item{...}{Le nom de la couleur.}
}
\value{
Un vecteur nomme avec les couleurs
}
\description{
\code{sgdf_colors()} donne les couleurs avec les codes hexadecimaux.
}
