% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged_template.R
\name{sgdf_paged_long}
\alias{sgdf_paged_long}
\title{Rapport PDF - long format}
\usage{
sgdf_paged_long(toc = FALSE, picto)
}
\arguments{
\item{toc}{Table des matieres}

\item{picto}{Picto a inclure en premiere page}
}
\value{
Un format Rmd sur mesure
}
\description{
Rapport PDF - long format
}
