% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged_template.R
\name{sgdf_paged_short}
\alias{sgdf_paged_short}
\title{Rapport PDF - format note courte}
\usage{
sgdf_paged_short(picto)
}
\arguments{
\item{picto}{Picto a inclure en premiere page}
}
\value{
Un format Rmd sur mesure
}
\description{
Rapport PDF - format note courte
}
