% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{sgdf_pal_inter}
\alias{sgdf_pal_inter}
\title{Interpolation des palettes}
\usage{
sgdf_pal_inter(palette = "pal_sgdf_main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{L'une des palettes de \link{sgdf_palettes}.}

\item{reverse}{Indique si il faut inverser la palette.}

\item{...}{Arguments de \code{\link[grDevices:colorRampPalette]{grDevices::colorRampPalette()}}.}
}
\value{
Renvoie un numeric (nb de couleurs) et un vecteur de couleurs.
}
\description{
Interpolation des palettes
}
\keyword{internal}
