% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{sgdf_palettes}
\alias{sgdf_palettes}
\title{Couleurs SGDF}
\format{
Une liste nommee des couleurs de la charte SGDF
}
\usage{
sgdf_palettes
}
\description{
Couleurs SGDF
}
\keyword{datasets}
