% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{theme_flex_sgdf}
\alias{theme_flex_sgdf}
\title{Theme SGDF pour les tableaux}
\usage{
theme_flex_sgdf(flex_table)
}
\arguments{
\item{flex_table}{Un tableau fait avec flextable()}
}
\value{
Un tableau flextable() mis en forme
}
\description{
Theme SGDF pour les tableaux
}
