% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_sgdf}
\alias{theme_sgdf}
\title{Theme ggplot2 pour les graphiques - Scouts et Guides de France}
\usage{
theme_sgdf(
  base_family = "Sarabun",
  head_family = "Caveat Brush",
  base_size = 12,
  plot_title_size = 20,
  subtitle_size = 16,
  caption_size = 10,
  text_col = "#003a5d",
  base_line_size = 0.8,
  main_background_color = "#FFFFFF",
  grid = "xy"
)
}
\arguments{
\item{base_family}{Police normale}

\item{head_family}{Police titres}

\item{base_size}{Taille normale}

\item{plot_title_size}{Taille titres}

\item{subtitle_size}{Taille sous-titre}

\item{caption_size}{Taille caption}

\item{text_col}{Couleur texte}

\item{base_line_size}{Taille lignes des grilles}

\item{main_background_color}{Couleur arriere-plan}

\item{grid}{Grille (no/x/y/xy)}
}
\value{
Un \code{ggplot2} theme.
}
\description{
Theme ggplot2 pour les graphiques - Scouts et Guides de France
}
