% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_sgdf_carte}
\alias{theme_sgdf_carte}
\title{Theme ggplot2 pour les cartes - Scouts et Guides de France}
\usage{
theme_sgdf_carte(
  base_family = "Sarabun",
  head_family = "Caveat Brush",
  base_size = 12,
  plot_title_size = 20,
  subtitle_size = 16,
  caption_size = 10,
  text_col = "#003a5d",
  main_background_color = "#FFFFFF"
)
}
\arguments{
\item{base_family}{Police normale}

\item{head_family}{Police titres}

\item{base_size}{Taille normale}

\item{plot_title_size}{Taille titres}

\item{subtitle_size}{Taille sous-titre}

\item{caption_size}{Taille caption}

\item{text_col}{Couleur texte}

\item{main_background_color}{Couleur arriere-plan}
}
\value{
Un \code{ggplot2} theme.
}
\description{
Theme ggplot2 pour les cartes - Scouts et Guides de France
}
\details{
Ce theme est a utiliser avec \code{ggplot2::geom_sf()}.
}
